Move objects logically belonging to function space into correct module
