+++ optimism/FunctionSpace.py
@@ -1,5 +1,7 @@
 from jax.scipy.linalg import solve
 from jax.lax import scan
+import numpy as onp
+
 from optimism.JaxConfig import *
 from optimism import Interpolants
 from optimism import Mesh
@@ -9,6 +11,7 @@ from optimism.TensorMath import tensor_2D_to_3D
 
 FunctionSpace = namedtuple('FunctionSpace', ['shapes', 'vols', 'shapeGrads', 'mesh', 'quadratureRule'])
 
+EssentialBC = namedtuple('EssentialBC', ['nodeSet', 'component'])
 
 def construct_function_space(mesh, quadratureRule, mode2D='cartesian'):
     shapes = vmap(compute_shape_values_on_element,
@@ -186,5 +189,91 @@ def average_quadrature_field_over_element(elemQPData, vols):
     S = np.tensordot(vols, elemQPData, axes=[0,0])
     elVol = np.sum(vols)
     return S/elVol
-    
 
+
+class DofManager:
+    def __init__(self, functionSpace, dim, EssentialBCs):
+        self.fieldShape = Mesh.num_nodes(functionSpace.mesh), dim
+        self.isBc = onp.full(self.fieldShape, False, dtype=bool)
+        for ebc in EssentialBCs:
+            self.isBc[functionSpace.mesh.nodeSets[ebc.nodeSet], ebc.component] = True
+        self.isUnknown = ~self.isBc
+
+        self.ids = np.arange(self.isBc.size).reshape(self.fieldShape)
+
+        self.unknownIndices = self.ids[self.isUnknown]
+        self.bcIndices = self.ids[self.isBc]
+
+        ones = np.ones(self.isBc.size, dtype=int) * -1
+        self.dofToUnknown = ones.at[self.unknownIndices].set(np.arange(self.unknownIndices.size)) 
+
+        self.HessRowCoords, self.HessColCoords = self._make_hessian_coordinates(functionSpace.mesh.conns)
+
+        self.hessian_bc_mask = self._make_hessian_bc_mask(functionSpace.mesh.conns)
+
+
+    def get_bc_size(self):
+        return np.sum(self.isBc).item() # item() method casts to Python int
+
+
+    def get_unknown_size(self):
+        return np.sum(self.isUnknown).item() # item() method casts to Python int
+
+
+    def create_field(self, Uu, Ubc=0.0):
+        U = np.zeros(self.isBc.shape).at[self.isBc].set(Ubc)
+        return U.at[self.isUnknown].set(Uu)
+
+
+    def get_bc_values(self, U):
+        return U[self.isBc]
+
+
+    def get_unknown_values(self, U):
+        return U[self.isUnknown]
+
+
+    def slice_unknowns_with_dof_indices(self, Uu, dofIndexSlice):
+        i = self.isUnknown[dofIndexSlice]
+        j = self.dofToUnknown.reshape(self.fieldShape)[dofIndexSlice]
+        return Uu[j[i]]
+
+
+    def _make_hessian_coordinates(self, conns):
+        nElUnknowns = onp.zeros(conns.shape[0], dtype=int)
+        nHessianEntries = 0
+        for e, eNodes in enumerate(conns):
+            elUnknownFlags = self.isUnknown[eNodes,:].ravel()
+            nElUnknowns[e] = onp.sum(elUnknownFlags)
+            nHessianEntries += onp.square(nElUnknowns[e])
+
+        rowCoords = onp.zeros(nHessianEntries, dtype=int)
+        colCoords = rowCoords.copy()
+        rangeBegin = 0
+        for e,eNodes in enumerate(conns):
+            elDofs = self.ids[eNodes,:]
+            elUnknownFlags = self.isUnknown[eNodes,:]
+            elUnknowns = self.dofToUnknown[elDofs[elUnknownFlags]]
+            elHessCoords = onp.tile(elUnknowns, (nElUnknowns[e],1))
+
+            rangeEnd = rangeBegin + onp.square(nElUnknowns[e])
+
+            rowCoords[rangeBegin:rangeEnd] = elHessCoords.ravel()
+            colCoords[rangeBegin:rangeEnd] = elHessCoords.T.ravel()
+
+            rangeBegin += np.square(nElUnknowns[e])
+        return rowCoords, colCoords
+
+
+    def _make_hessian_bc_mask(self, conns):
+        nElements, nNodesPerElement = conns.shape
+        nFields = self.ids.shape[1]
+        nDofPerElement = nNodesPerElement*nFields
+
+        hessian_bc_mask = onp.full((nElements,nDofPerElement,nDofPerElement),
+                                   True, dtype=bool)
+        for e, eNodes in enumerate(conns):
+            eFlag = self.isBc[eNodes,:].ravel()
+            hessian_bc_mask[e,eFlag,:] = False
+            hessian_bc_mask[e,:,eFlag] = False
+        return hessian_bc_mask
